Various fixed to remove unused variables and compiler type conversion warnings by Ralph Campbell

+++ src/client/gui/shortcutwindow.h
@@ -142,7 +142,7 @@ protected:
 
     csString buttonBackgroundImage;
 
-    int edit;
+    size_t edit;
     pawsWidget *editedButton;
 
     virtual void HandleMessage(MsgEntry *msg);
